Simplify dashboard implmenetation; remove cruft
