+++ packages/esm-patient-chart-app/src/view-components/chart-review.scss
@@ -7,3 +7,32 @@
   margin: 1rem 0rem 1rem 1.3125rem;
   color: $text-01;
 }
+
+.dashboard {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  grid-auto-rows: auto;
+  grid-auto-columns: auto;
+  grid-gap: 25px;
+  margin: 1.3125rem;
+}
+
+:global(.omrs-breakpoint-lt-desktop) .dashboard {
+  grid-template-columns: 1fr;
+}
+
+:global(.omrs-breakpoint-lt-tablet) .container {
+  margin: 5px;
+  overflow-x: auto;
+  justify-content: center;
+}
+
+:global(.omrs-breakpoint-lt-tablet) .dashboard {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: wrap;
+}
+
+:global(.omrs-breakpoint-lt-tablet) .dashboard > div {
+  margin: 2px 0px;
+}
