--- packages/esm-patient-chart-app/src/view-components/tabbed-view.css
@@ -1,69 +1,0 @@
-.summariesnav {
-  display: flex;
-  position: sticky;
-  top: 3rem;
-  flex-direction: row;
-  align-items: left;
-  justify-content: flex-start;
-  width: 100%;
-  background-color: var(--omrs-color-bg-low-contrast);
-  height: 3.125rem;
-  padding: 0 1.3125rem;
-  flex-wrap: nowrap;
-  -webkit-overflow-scrolling: touch;
-  border-bottom: 1px solid var(--omrs-color-bg-lowest-contrast);
-}
-
-.summariesnav::-webkit-scrollbar {
-  display: none;
-}
-
-.summariesnav ul {
-  flex: 0 0 auto;
-  margin-block-start: 0;
-  margin-block-end: 0;
-  margin-inline-start: 0;
-  margin-inline-end: 0;
-  padding-inline-start: 0;
-}
-
-.summariesnav ul li {
-  float: left;
-  list-style-type: none;
-  text-align: center;
-}
-
-.summariesnav button {
-  padding: 0.9rem;
-  cursor: pointer;
-  border: none;
-}
-
-.summariesnav button:focus {
-  outline: none !important;
-}
-
-.summariesnav ul li :global(div.selected) {
-  color: var(--omrs-color-ink-high-contrast);
-  background-color: var(--omrs-color-bg-low-contrast);
-  border-bottom: medium solid var(--omrs-color-ink-high-contrast);
-
-  border-radius: 1.5px 1.5px 0px 0px;
-  font-weight: 500;
-}
-
-.summariesnav ul li :gobal(div.unselected) {
-  color: var(--omrs-color-ink-medium-contrast);
-}
-
-.summariesnav a {
-  color: inherit;
-}
-
-.routesContainer {
-  margin: 1.3125rem;
-}
-
-:global(.omrs-breakpoint-lt-tablet) .routesContainer {
-  margin: 0.625rem;
-}
